engine/requirements.txt
fastapi
uvicorn
requests
openai
python-multipart
PyPDF2
pillow
pytesseract
python-dotenv
pdf2image
pgvector
psycopg2-binary
sqlalchemy
typing-extensions
langchain-text-splitters
tiktoken
sentence-transformers
ollama
torch
numpy
celery[redis]
redis
